last changes

+++ src/components/team/team.tsx
@@ -9,12 +9,11 @@ import { updateTeam } from "../../redux/tems-reducer/team-actions";
 import { teamSlice } from "../../redux/tems-reducer/teams-reducer";
 
 const Team: React.FC<{
-
   leadName: string;
   employees: IEmployee[];
   id?: number;
-
 }> = ({ leadName, employees, id }) => {
+
   const dispatch = useAppDispatch();
   const { addName, removeTeam } = teamSlice.actions;
 
@@ -24,12 +23,10 @@ const Team: React.FC<{
 
   const handleSubmit = (event: React.FormEvent) => {
     event.preventDefault();
-    
-    if (name.trim() !== "") {
 
+    if (name.trim() !== "") {
       dispatch(addName({ leadName, name, id }));
       setTimeout(() => newletter(""), 100);
-      
     } else {
       alert("Введите имя сотрудника");
     }
@@ -40,10 +37,14 @@ const Team: React.FC<{
   };
 
   const removeClick = (event: React.FormEvent) => {
+    
     event.stopPropagation();
-
-    dispatch(removeTeam({ id }));
-    dispatch(updateTeam(id!));
+    const areUShure = window.confirm(`Удалить команду ${leadName}?`);
+    if (areUShure) {
+      dispatch(removeTeam({ id }));
+      dispatch(updateTeam(id!));
+    }
+    
   };
 
   return (
@@ -86,7 +87,13 @@ const Team: React.FC<{
           <input type="submit" value="+" className="teams-add small " />
         </form>
         {employees.map((i, indx) => (
-          <Employee key={indx} id={id!} leadName={leadName} name={i.name!} indx={indx} />
+          <Employee
+            key={indx}
+            id={id!}
+            leadName={leadName}
+            name={i.name!}
+            indx={indx}
+          />
         ))}
       </div>
     </div>
